[REFACTOR(Hand)] 패키지변경
